Relocate existing input controls
